Move views into folder
